requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
boto3==1.18.16
botocore==1.21.17
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-bootstrap-form==3.4
django-filter==2.4.0
django-location-field==2.1.0
django-storages==1.11.1
django-tables2==2.3.4
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.0
paramiko==2.7.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
storage==0.0.4.3
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
